moved firebase into a seperate file
