+++ firebase/index.js
@@ -1,6 +1,20 @@
 import * as firebase from 'firebase';
 
+var config = {
+    apiKey: "AIzaSyDmqsfVyAB0rmYpms3YczKsrFPrYTXuyes",
+    authDomain: "reactapp-a97ee.firebaseapp.com",
+    databaseURL: "https://reactapp-a97ee.firebaseio.com",
+    storageBucket: "reactapp-a97ee.appspot.com",
+};
+
+const firebaseApp = firebase.initializeApp(config);
+
+export {firebaseApp}
 
 function createUser (name, description, location, story, image, badges) {
+
+}
+
+function createMessage (message, to, from) {
 
 }
