+++ index.ios.js
@@ -12,15 +12,7 @@ import {
   View
 } from 'react-native';
 import Test from './components/testComp.js'
-import * as firebase from 'firebase';
-var config = {
-    apiKey: "AIzaSyDmqsfVyAB0rmYpms3YczKsrFPrYTXuyes",
-    authDomain: "reactapp-a97ee.firebaseapp.com",
-    databaseURL: "https://reactapp-a97ee.firebaseio.com",
-    storageBucket: "reactapp-a97ee.appspot.com",
-};
-
-const firebaseApp = firebase.initializeApp(config);
+import {firebaseApp} from './firebase'
